Create EnRonEmail_Network.py
This file contains the code for my project on Enron Email dataset. I tried multiple layers of data cleaning on Enron dataset because it was a large 1+ GB of text data, all attached together in one single file. After cleaning, I used the email response time of every employee to infer their role in the company.